Make SpaceShip class to protocol
